fix(web): improve type safety

+++ web/src/lib/components/Modals/DeletePartitionsModal.svelte
@@ -49,6 +49,7 @@
 
       async onUpdate({ form }) {
         if (!form.valid) return;
+        if (!page.params.streamId || !page.params.topicId) return;
 
         const { data, ok } = await fetchRouteApi({
           method: 'DELETE',
